// src/app/api/generated/model/create-precio-particular-request.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Precio particular de la oferta para un periodo
 */
export interface CreatePrecioParticularRequest { 
    /**
     * Precio particular en ese periodo. Cero es valido
     */
    importe: number;
    /**
     * ISO 4217. Si se omite, la del precio de lista de la oferta
     */
    moneda?: string;
    /**
     * Primer dia en que rige
     */
    vigenciaDesde: string;
    /**
     * ULTIMO dia en que rige, INCLUSIVE. Null = sin fin previsto
     */
    vigenciaHasta?: string;
}